コンポーネントの細分化
