+++ front/src/templates/PrefectureDailyDead.tsx
@@ -53,6 +53,13 @@ export const PrefectureDailyDead = () => {
               累計死者数
             </ReChart>
           )}
+          <a
+           href="https://www3.nhk.or.jp/news/special/coronavirus/data-widget/"
+           target="_blank"
+           rel="noopener"
+            >
+           都道府県ごとの死者数の推移情報提供:NHK
+             </a>
         </Inner>
       )}
     </>
